feature/admin: able to create and display requests
