+++ src/pages/admin/index.tsx
@@ -1,65 +1,11 @@
+import DisplayRequests from "@/components/Admin/DisplayRequests";
+import FormRequest from "@/components/FormRequest";
 import MainLayout from "@/components/layouts/MainLayout";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { useForm } from "react-hook-form";
-import { type z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Popover,
-  PopoverTrigger,
-  PopoverContent,
-} from "@/components/ui/popover";
-import { cn } from "@/utils/utils";
-import { format } from "date-fns";
-import { Calendar } from "@/components/ui/calendar";
-import { CalendarIcon } from "lucide-react";
-import { Textarea } from "@/components/ui/textarea";
-import { Badge } from "@/components/ui/badge";
-import { insertRequestSchema } from "@/server/db/schema";
 import { useSession } from "next-auth/react";
 import router from "next/router";
-import { useEffect, useState } from "react";
-import { api } from "@/utils/api";
-import { toast } from "@/components/ui/use-toast";
+import { useEffect } from "react";
 
 export default function Page() {
-  const [open, setOpen] = useState(false);
-
-  const form = useForm<z.infer<typeof insertRequestSchema>>({
-    resolver: zodResolver(insertRequestSchema),
-    defaultValues: {
-      maxTotalPrice: 1,
-      location: "",
-      numGuests: 1,
-      minNumBeds: 1,
-      minNumBedrooms: 1,
-      note: "",
-    },
-  });
-
   const { data } = useSession();
 
   // Only admin can access this page
@@ -73,295 +19,14 @@ export default function Page() {
     return null;
   }
 
-  const utils = api.useUtils(); // To allow to invalidate the data useContext depracated
-
-  const { mutate, isLoading } = api.requests.create.useMutation({
-    onSuccess: () => {
-      setOpen(false);
-      void utils.requests.getAll.invalidate(); // will revalidate the tasks array to see if there are any changes
-    },
-    onError: (error) => {
-      toast({
-        title: "Something went wrong",
-        description: error.message,
-      });
-    },
-  });
-
-  function onSubmit(values: z.infer<typeof insertRequestSchema>) {
-    mutate(values);
-  }
-
   return (
     <MainLayout pageTitle="Admin Dashboard">
-      <main className="container">
-        <p className="p-8">admin dashboard</p>
-        <Dialog open={open} onOpenChange={setOpen}>
-          <DialogTrigger asChild>
-            <Button>+ Create a request</Button>
-          </DialogTrigger>
-          <DialogContent>
-            <DialogHeader>
-              <DialogTitle>
-                Request
-                <Badge
-                  variant="secondary"
-                  size="sm"
-                  className="ml-2 -translate-y-0.5 uppercase"
-                >
-                  Admin
-                </Badge>
-              </DialogTitle>
-              <DialogDescription>Add personal request</DialogDescription>
-            </DialogHeader>
-
-            {/* Ensure that the Form component wraps the relevant content */}
-            <Form {...form}>
-              <form
-                onSubmit={form.handleSubmit(onSubmit)}
-                className="mt-5 space-y-5"
-              >
-                <section className="flex flex-row gap-5">
-                  <FormField
-                    control={form.control}
-                    name="maxTotalPrice"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Max Price</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="1"
-                            type={"number"}
-                            className="w-[100px]"
-                            {...field}
-                            onChange={(e) => {
-                              field.onChange(parseInt(e.target.value));
-                            }}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="location"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Location</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="Dubai"
-                            type="text"
-                            className="w-full"
-                            {...field}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </section>
-
-                <section className="flex flex-row gap-5">
-                  <FormField
-                    control={form.control}
-                    name="checkIn"
-                    render={({ field }) => (
-                      <FormItem className="flex flex-col">
-                        <FormLabel>Check-in date</FormLabel>
-                        <Popover>
-                          <PopoverTrigger asChild>
-                            <FormControl>
-                              <Button
-                                variant={"outline"}
-                                className={cn(
-                                  "w-[200px] pl-3 text-left font-normal",
-                                  !field.value && "text-muted-foreground",
-                                )}
-                              >
-                                {field.value ? (
-                                  format(field.value, "PPP")
-                                ) : (
-                                  <span>Pick a date</span>
-                                )}
-                                <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
-                              </Button>
-                            </FormControl>
-                          </PopoverTrigger>
-                          <PopoverContent className="w-auto p-0" align="start">
-                            <Calendar
-                              mode="single"
-                              selected={new Date(field.value)}
-                              onSelect={(date) =>
-                                field.onChange(date?.toISOString())
-                              } // Convert date to ISO string
-                              disabled={(date) => date < new Date()}
-                            />
-                          </PopoverContent>
-                        </Popover>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="checkOut"
-                    render={({ field }) => (
-                      <FormItem className="flex flex-col">
-                        <FormLabel>Check-in out</FormLabel>
-                        <Popover>
-                          <PopoverTrigger asChild>
-                            <FormControl>
-                              <Button
-                                variant={"outline"}
-                                className={cn(
-                                  "w-[200px] pl-3 text-left font-normal",
-                                  !field.value && "text-muted-foreground",
-                                )}
-                              >
-                                {field.value ? (
-                                  format(field.value, "PPP")
-                                ) : (
-                                  <span>Pick a date</span>
-                                )}
-                                <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
-                              </Button>
-                            </FormControl>
-                          </PopoverTrigger>
-                          <PopoverContent className="w-auto p-0" align="start">
-                            <Calendar
-                              mode="single"
-                              selected={new Date(field.value)}
-                              onSelect={(date) =>
-                                field.onChange(date?.toISOString())
-                              } // Convert date to ISO string
-                              disabled={(date) => date < new Date()}
-                            />
-                          </PopoverContent>
-                        </Popover>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </section>
-
-                <section className="flex flex-row gap-5">
-                  <FormField
-                    control={form.control}
-                    name="numGuests"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Guests</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="1"
-                            type={"number"}
-                            min={1}
-                            {...field}
-                            onChange={(e) => {
-                              field.onChange(parseInt(e.target.value));
-                            }}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="minNumBeds"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Beds</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="1"
-                            type={"number"}
-                            min={1}
-                            value={field.value?.toString()}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="minNumBedrooms"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Bedrooms</FormLabel>
-                        <FormControl>
-                          <Input
-                            placeholder="1"
-                            type={"number"}
-                            min={1}
-                            value={field.value?.toString()}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </section>
-
-                <FormField
-                  control={form.control}
-                  name="propertyType"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Property Type</FormLabel>
-                      <FormControl>
-                        <Select onValueChange={field.onChange}>
-                          <SelectTrigger className="w-[180px]">
-                            <SelectValue placeholder="Select a type" />
-                          </SelectTrigger>
-                          <SelectContent>
-                            <SelectItem value="townhouse">Townhouse</SelectItem>
-                            <SelectItem value="apartment">Apartment</SelectItem>
-                            <SelectItem value="house">House</SelectItem>
-                            <SelectItem value="guesthouse">
-                              Guest House
-                            </SelectItem>
-                          </SelectContent>
-                        </Select>
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-
-                <FormField
-                  control={form.control}
-                  name="note"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Note</FormLabel>
-                      <FormControl>
-                        <Textarea
-                          placeholder="Any notes about the property"
-                          className="resize-none"
-                          onChange={(e) => field.onChange(e.target.value)}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-
-                <Button type="submit" isLoading={isLoading}>
-                  Submit
-                </Button>
-              </form>
-            </Form>
-          </DialogContent>
-        </Dialog>
+      <main className="container flex flex-col space-y-5">
+        <h1 className="text-bold text-4xl uppercase">admin dashboard</h1>
+        <div>
+          <FormRequest />
+        </div>
+        <DisplayRequests />
       </main>
     </MainLayout>
   );
